initial commit. A basic working version of bills/invoices.
The successor project developed by Martin Klampfer was called
"invoices". To prevent name conflicts this project runs for now under
the name "bills".  Fortunately this also prevents a conflict with
index.* at tab completition.  The project name might be changed
again at a later point in time.

Currently the project uses the original database. But the database
schema might get changed soon, reducing it to the actual requirements
and removing a few irregularities.