removed the $default_callback callable from HtmlFormOptions class
the callable can not be used as a fallback for non defined callables
in the constructor, since the the given type 'callable' does not allow
for None as a valid type.
The implementation of function/method calls in PHP are not satisfactory:
they do not allow named parameters, and no composite types.