fixed/added database insert tests
The problem with test_insert_lineitem() was that instead of the required
`price` field accidentally a `value` field was defined. It is a sad fact
that there is no obvious feedback in such a failed query does not lead
to any obvious feedback or error message except for the missing entry
in the database. (Of course checking checking pdo messages might me an
idea.) Ultimately checking consistency of the data with the sql schema
already in the Record at data input or at least at handing the data over
to the database methods might be the only sustainable solution.